Override slug as well so that all pages have unique slugs
Pages with identical base filenames (but different paths) result in
identical slugs by default, which fails uniqueness and interferes with
the translation logic.

We now override the slug to be equal to the full relative path to
ensure it is unique.

Overriding the slug needs to take place earlier to prevent
interference with the translation logic, so we use 2 signals:

1. content_object_init: override metadata (url, slug, save_as). This
   signal provides access to a single page at a time, so we cannot set
   the parent/child relationships here.
2. page_generator_finalized: set page parent/child relationships.