Support for non-forced basename slugs and translations
The URL and SAVE_AS overrides are now generated from
PAGE_URL and PAGE_SAVE_AS (and PAGE_LANG_URL and PAGE_LANG_SAVE_AS).
To rid of 'pages/' prefix, PATH_METADATA is no longer needed if
PAGE_URL is set to exclude it. Also slugs can be slugified from
the title or basename or set in the header. Translations should
also work all as expected.

When there is no parent for a translated page, the parent of the
non-translated equivalent is used.

Closes #1.